functions/requirements.txt

firebase_functions~=0.1.0
requests
dataflows
pyproj
python-dotenv
kvfile
python-slugify
flask
